break out serialization.ts from metamodel.ts
